+++ main-src/eu/ha3/matmos/game/data/MAtDataGatherer.java
@@ -5,7 +5,7 @@ import java.util.Locale;
 import java.util.Set;
 
 import net.minecraft.src.EntityPlayer;
-import net.minecraft.src.Ha3Utility;
+import eu.ha3.mc.haddon.implem.Ha3Utility;
 import net.minecraft.src.ItemStack;
 import net.minecraft.src.Minecraft;
 import net.minecraft.src.PotionEffect;
