+++ main-src/eu/ha3/matmos/game/user/MAtUserControl.java
@@ -1,7 +1,7 @@
 package eu.ha3.matmos.game.user;
 
 import net.minecraft.src.GuiScreen;
-import net.minecraft.src.Ha3Utility;
+import eu.ha3.mc.haddon.implem.Ha3Utility;
 import net.minecraft.src.KeyBinding;
 import net.minecraft.src.Minecraft;
